[RESOLUTION][INT]: support int for cam resolution

+++ include/cv_camera/driver.h
@@ -135,7 +135,6 @@ class Driver : public rclcpp::Node
    * @brief Video path
   */
   std::string video_path_;
-
   /**
    * @brief Camera info topic
   */
@@ -170,11 +169,11 @@ class Driver : public rclcpp::Node
   /**
    * @brief Camera cv_cap_prop_frame_width.
    */
-  float width_;
+  int width_;
   /**
    * @brief Camera cv_cap_prop_frame_height.
    */
-  float height_;
+  int height_;
   /**
    * @brief Camera cv_cap_prop_brightness.
    */
